+++ lib/ui/events2/Event2DetailPanel.dart
@@ -223,7 +223,7 @@ class _Event2DetailPanelState extends Event2Selector2State<Event2DetailPanel> im
     Text(_displayCategories?.join(', ') ?? '', overflow: TextOverflow.ellipsis, maxLines: 2, style: Styles().textStyles.getTextStyle("widget.card.title.small.fat"));
 
   static List<String>? _buildDisplayCategories(Event2? event) =>
-    Events2().contentAttributes?.displaySelectedLabelsFromSelection(event?.attributes, usage: ContentAttributeUsage.category);
+    Events2().displaySelectedContentAttributeLabelsFromSelection(event?.attributes, usage: ContentAttributeUsage.category);
 
   Widget get _groupingBadgeWidget {
     String? badgeLabel;